+++ react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.0.2 (July 28, 2021)
+
+### Fixed
+
+-   Issue with undefined flexDirection causing crash on Android.
+-   Incorrect fill alignment and color for UPS icon.
+
 ## v1.0.1 (July 15, 2021)
 
 ### Added